Правка media_clear

+++ commands/media_clear_command.py
@@ -1,3 +1,4 @@
+from disnake import MessageType
 from bot_init import bot
 from commands.misc.check_roles import has_any_role_by_id
 from config import HEAD_ADT_TEAM
@@ -15,16 +16,16 @@ async def media_clear(ctx, count: int):
         deleted = 0
         async for message in ctx.channel.history(limit=count):
             # Пропускаем системные сообщения
-            if message.type != message.Type.default:
+            if message.type != MessageType.default:
                 continue
-            
+
             # Проверяем, содержит ли сообщение вложения (медиа, файлы и т. д.)
-            if not message.attachments:
+            if not message.attachments and not message.content.startswith("https://"):
                 await message.delete()
                 deleted += 1
-        
+
         await ctx.send(f"✅ Удалено {deleted} сообщений без медиафайлов.", delete_after=5)
 
     except Exception as e:
         print(f"❌ Ошибка при удалении сообщений: {e}")
-        await ctx.send("❌ Произошла ошибка при удалении сообщений. Проверьте логи.")
+        await ctx.send(f"❌ Произошла ошибка при удалении сообщения: {message} Проверьте логи.")
